converttrajectory 1to3 enhancement final

+++ crs-converter-core/src/main/java/org/opengroup/osdu/crs/model/v4/MinimumDepthInterval.java
@@ -21,6 +21,6 @@ public class MinimumDepthInterval {
     @ApiModelProperty(
             value = Constants.SWAGGER_MD_INTERVAL,
             example = Constants.SWAGGER_MD_INTERVAL_EXAMPLE)
-    private Integer md_interval;
+    private Double md_interval;
 
 }
